+++ tests/testthat/test_cluster_events.R
@@ -1,13 +1,13 @@
 context("Cluster functions")
 
-exclude_dbscan <- TRUE
+exclude_dbscan <- FALSE
 
 # Preparation -------------------------------------------------------------
-library(testthat)
 library(beadplexr)
 data("lplex")
 .data <- lplex[[1]]
 .parameters <- c("FSC-A", "SSC-A")
+.panel_info <- load_panel(.panel_name = "Human Growth Factor Panel (13-plex)")
 
 # Test bp_kmeans ----------------------------------------------------------
 test_that("bp_kmeans works", {
@@ -58,11 +58,19 @@ test_that("bp_clara fails", {
 })
 
 # Test bp_dbscan ----------------------------------------------------------
+bp_dbscan_num_clusters <- function(.data, .parameter, .eps, .MinPts, scale){
+  dbc <- bp_dbscan(.data, .parameter = .parameter, .eps = .eps, .MinPts = .MinPts, .column_name = "cluster", scale = scale)
+  dbc$cluster <- as.numeric(dbc$cluster)
+  max(dbc$cluster, na.rm = TRUE)
+}
+
+
 test_that("bp_dbscan works", {
   # Defaults
   expect_is(bp_dbscan(.data = .data, .parameter = .parameters, .column_name = "clu ster"), "data.frame")
   if(exclude_dbscan){
-    skip("Indepth test of bp_dbscan")
+    skip("In-depth test of bp_dbscan")
+  }else{
     # Set eps and MinPts
     expect_is(bp_dbscan(.data = .data, .parameter = .parameters, .column_name = "cluster", .eps = 0.1), "data.frame")
     expect_is(bp_dbscan(.data = .data, .parameter = .parameters, .column_name = "cluster", .MinPts = 100), "data.frame")
@@ -73,6 +81,18 @@ test_that("bp_dbscan works", {
     expect_is(bp_dbscan(.data = .data, .parameter = .parameters, .column_name = "cluster", method = "raw"), "data.frame")
     expect_is(bp_dbscan(.data = .data, .parameter = .parameters, .column_name = "cluster", seeds = FALSE), "data.frame")
     expect_is(bp_dbscan(.data = .data, .parameter = .parameters, .column_name = "cluster", scale = TRUE), "data.frame")
+
+    # Problematic dbscan
+    .analyte_set <- bp_dbscan(.data = .data, .parameter = .parameters, .column_name = "bead_cluster", scale = TRUE)
+
+    .bead_a <- dplyr::filter(.analyte_set, bead_cluster == "2")
+    .bead_b <- dplyr::filter(.analyte_set, bead_cluster == "1")
+
+    expect_equal(bp_dbscan_num_clusters(.data = .bead_a, .parameter = c("FL6-H", "FL2-H"), .eps = 0.09, .MinPts = 50, scale = FALSE), 6)
+    expect_equal(bp_dbscan_num_clusters(.data = .bead_a, .parameter = c("FL6-H", "FL2-H"), .eps = 0.12, .MinPts = 100, scale = FALSE), 6)
+
+    expect_equal(bp_dbscan_num_clusters(.data = .bead_b, .parameter = c("FL6-H", "FL2-H"), .eps = 0.095, .MinPts = 50, scale = FALSE), 7)
+    expect_equal(bp_dbscan_num_clusters(.data = .bead_b, .parameter = c("FL6-H", "FL2-H"), .eps = 0.2, .MinPts = 110, scale = FALSE), 7)
   }
 })