+++ NEWS.md
@@ -1,3 +1,16 @@
+# beadplexr version 0.2.0
+
+## Minor changes
+
+  + Add BioLegend panel information
+    + Human Adipokine Panel* (13-plex)
+    + Human Anti-Virus Response Panel (13-plex)
+    + Human CD8/NK Panel (13-plex)
+    + Human Inflammation Panel (13-plex)
+    + Human Metabolic Panel 1 (4-plex)
+  + Add citation
+  + Fix problem of missing column names after upgrade of `tibble` to 2.0
+
 # beadplexr version 0.1.0
 
 Initial release - clip the flag!